parser-worker/requirements.txt
supabase>=2.22.3,<3
pdfplumber==0.11.7
pypdf==6.1.3
Pillow==11.3.0
python-dotenv==1.1.1
pytest==8.4.2
